chore: move generateBound
